+++ .eslintrc
@@ -10,6 +10,7 @@
   },
   "rules": {
     "no-underscore-dangle": ["error", { "allow": ["_id"] }],
-    "no-console": 0
+    "no-console": 0,
+    "linebreak-style": 0
   }
 }
