+++ app.js
@@ -1,90 +1,32 @@
-require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const helmet = require('helmet');
 const cors = require('cors');
-const { errors, celebrate, Joi } = require('celebrate');
-
-const usersRouter = require('./routes/users');
-const cardsRouter = require('./routes/movies');
-
-const NotFoundError = require('./errors/not-found-err');
-
-const { addUser, login } = require('./controllers/users');
-const auth = require('./middlewares/auth');
-// const urlvalidator = require('./middlewares/url-validation');
+const { errors } = require('celebrate');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-
-const { PORT = 8000 } = process.env;
-const { MONGO_URI = 'mongodb://localhost:27017/bitfilmsdb' } = process.env;
-
-console.dir(MONGO_URI);
+const router = require('./routes');
+const errorhandler = require('./middlewares/error-handler');
+const { PORT, MONGO_URI, MONGO_OPTIONS } = require('./config');
 
 const app = express();
+
 app.use(cors());
 
-mongoose.connect(MONGO_URI, {
-  useNewUrlParser: true,
-  useCreateIndex: true,
-  useFindAndModify: false,
-  useUnifiedTopology: true,
-});
+mongoose.connect(MONGO_URI, MONGO_OPTIONS);
+
+app.use(requestLogger);
 
 app.use(helmet());
 
 app.use('/', express.json());
 
-app.use(requestLogger);
-
-// эмуляция падения серевера
-app.get('/crash-test', () => {
-  setTimeout(() => {
-    throw new Error('Сервер сейчас упадёт');
-  }, 0);
-});
-
-app.post(
-  '/signin',
-  celebrate({
-    body: Joi.object().keys({
-      email: Joi.string().required().email(),
-      password: Joi.string().required().min(8),
-    }),
-  }),
-  login,
-);
-app.post(
-  '/signup',
-  celebrate({
-    body: Joi.object().keys({
-      email: Joi.string().required().email(),
-      password: Joi.string().required().min(8),
-      name: Joi.string().min(2).max(30),
-      // avatar: Joi.string().custom(urlvalidator, 'custom URL validator'),
-    }),
-  }),
-  addUser,
-);
-app.use(auth);
-app.use('/', usersRouter);
-app.use('/', cardsRouter);
-app.use('*', (req, res, next) => {
-  next(new NotFoundError('Запрашиваемый ресурс не найден'));
-});
+app.use(router);
 
 app.use(errorLogger);
 
 app.use(errors());
-// Обработчик ошибок
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-  res.status(statusCode).send({
-    message: statusCode === 500
-      ? 'На сервере произошла ошибка'
-      : message,
-  });
-  next();
-});
+
+app.use(errorhandler);
 
 app.listen(PORT, () => {
   console.log(`Server started at port: ${PORT}`);
